requirements.txt
Flask>=2.3.3
pandas>=2.1.0
numpy>=1.24.0
python-pptx==0.6.21
Werkzeug>=2.3.7